feat(SDK-4481) moves FCM constant to correct place
