+++ clevertap-core/src/main/java/com/clevertap/android/sdk/pushnotification/PushConstants.java
@@ -6,7 +6,7 @@ public interface PushConstants {
 
     String LOG_TAG = "PushProvider";
 
-    String FCM_LOG_TAG = PushType.FCM.toString();
+    String FCM_LOG_TAG = "FCM";
 
     @NonNull
     String FCM_DELIVERY_TYPE = "fcm";
@@ -15,4 +15,12 @@ public interface PushConstants {
     String FIREBASE_SDK_CLASS = "com.google.firebase.messaging.FirebaseMessagingService";
     String FCM_PROPERTY_REG_ID = "fcm_token";
     String FCM_PUSH_TYPE_LOG_NAME = "FCM";
+
+    PushType FCM = new PushType(
+            FCM_DELIVERY_TYPE,
+            FCM_PROPERTY_REG_ID,
+            CT_FIREBASE_PROVIDER_CLASS,
+            FIREBASE_SDK_CLASS,
+            FCM_PUSH_TYPE_LOG_NAME
+    );
 }
