+++ clevertap-core/src/main/java/com/clevertap/android/sdk/pushnotification/PushType.java
@@ -7,14 +7,6 @@ import org.json.JSONObject;
 
 public class PushType {
 
-    public static PushType FCM = new PushType(
-            PushConstants.FCM_DELIVERY_TYPE,
-            PushConstants.FCM_PROPERTY_REG_ID,
-            PushConstants.CT_FIREBASE_PROVIDER_CLASS,
-            PushConstants.FIREBASE_SDK_CLASS,
-            PushConstants.FCM_PUSH_TYPE_LOG_NAME
-    );
-
     private final String ctProviderClassName;
 
     private final String messagingSDKClassName;
@@ -55,14 +47,10 @@ public class PushType {
         return type;
     }
 
-    public String name() {
-        return logName;
-    }
-
     @NonNull
     @Override
     public String toString() {
-        return " [PushType:" + name() + "] ";
+        return " [PushType:" + logName + "] ";
     }
 
     // Method to convert PushType to JSONObject
