+++ clevertap-core/src/test/java/com/clevertap/android/sdk/pushnotification/fcm/FcmSdkHandlerImplTest.kt
@@ -2,7 +2,7 @@ package com.clevertap.android.sdk.pushnotification.fcm
 
 import com.clevertap.android.sdk.ManifestInfo
 import com.clevertap.android.sdk.pushnotification.CTPushProviderListener
-import com.clevertap.android.sdk.pushnotification.PushType.FCM
+import com.clevertap.android.sdk.pushnotification.PushConstants.FCM
 import com.clevertap.android.sdk.pushnotification.fcm.TestFcmConstants.Companion.FCM_SENDER_ID
 import com.clevertap.android.sdk.utils.PackageUtils
 import com.clevertap.android.shared.test.BaseTestCase
@@ -101,7 +101,9 @@ class FcmSdkHandlerImplTest : BaseTestCase() {
 
     @Test
     fun testGetPushType() {
-        Assert.assertEquals(handler.pushType, FCM)
+        Assert.assertEquals(handler.pushType,
+            FCM
+        )
     }
 
     /*@Test
@@ -141,7 +143,9 @@ class FcmSdkHandlerImplTest : BaseTestCase() {
         mockStatic(FirebaseMessaging::class.java).use {
             `when`(FirebaseMessaging.getInstance()).thenThrow(RuntimeException("Something Went wrong"))
             handler.requestToken()
-            verify(listener, times(1)).onNewToken(null, FCM)
+            verify(listener, times(1)).onNewToken(null,
+                FCM
+            )
         }
     }
 }
